Skripte in eigene Dateien ausgelagert für die übersichtlichkeit
